Fix: BWBCoverBot's tests no longer depend on the model's attribute order
Previously the tests would fail if an SQLModel's attributes returned in
a different order. E.g.:
```
- [(EditionCoverData(isbn_13='444', cover_exists=True),)]
+ [(EditionCoverData(cover_exists=True, isbn_13='444'),)]
```

Before comparison, the `EditionCoverData` models, from `SQLModel`, are now
represented as dictionaries so the order doesn't matter.

Additionally, the call to the deprecated `session.execute()` was
replaced with a call to `session.exec()`, and the tests were updated
to reflect model objects being returned directly, rather than being
nested in tuples.